Move history/event -> event
