+++ web/src/player/EditMenu.tsx
@@ -4,7 +4,7 @@ import theme from 'style/theme';
 import * as icons from 'style/icon';
 
 import * as Game from 'game';
-import * as Event from 'history/event';
+import * as Event from 'event';
 import * as server from 'server';
 import * as Player from 'player';
 import * as Stream from 'sseStream';
