+++ web/src/record/Roll.tsx
@@ -4,7 +4,7 @@ import * as dice from 'Dice';
 import * as humanTime from 'HumanTime';
 import * as Roll from './RollComponents';
 
-import * as Event from 'history/event';
+import * as Event from 'event';
 import * as Share from 'share';
 import * as rollStats from 'rollStats';
