+++ web/src/record/RollComponents.tsx
@@ -5,7 +5,7 @@ import * as dice from 'Dice';
 import * as icons from 'style/icon';
 
 import type { Connection } from 'connection';
-import * as Event from 'history/event';
+import * as Event from 'event';
 import * as Share from 'share';
 import * as rollStats from 'rollStats';
 import * as routes from 'routes';
